+++ tests/fixtures/noqa.py
@@ -520,3 +520,4 @@ CheckStopIteration = 1  # noqa: WPS440
 print(literal)  # noqa: WPS441
 
 unhashable = {[]}  # noqa: WPS442
+unhashable = [] * 2  # noqa: WPS443